Add a new method iterateWithContact
This handles the case that there are contact/mortar relations
between the continua.  Unfortunately, this does not quite fit
into the framework: so far we have treated all continua individually.
However, the contact solvers lumps everything into a single
algebraic system.  Consequently this is now getting a bit
hacky.  I hope to find something prettier in the future.

The linearized NtD map for continua is not implemented yet.

[[Imported from SVN: r6868]]