Add a very crude implementation of a 3x3 skew-symmetric matrix.
And use it to implement the Rotation<3> class.  This makes the
Rotation implementation easier to understand.  Previously we
have used a FieldVector<3> when a skew matrix was meant.
This leads to confusion now that we also use quaternions as
tangent vectors of SO(3).  In local coordinates, these also
have 3 entries.  Adding the new skew matrix class makes it
clearer what kind of mathematical object is meant.

[[Imported from SVN: r7897]]