Do not normalize unitvectors / unit quaternions in constructors and operator=
This effectively means that we use another prolongation of the distance 
function on M into the surrounding space.  Since the prolongation does not
matter this patch should not change the algorithm behavior.  However, it
shaves off a few norm calculations and division.  I cannot really measure
any difference though.

A possible effect of this is that while all values should remain on the
manifold, they may start to 'drift away' due to numerical artifacts.
So we may have to add an occasional renormalization step eventually.

[[Imported from SVN: r9558]]