Rewrite the exp method with embedded tangent vector argument
The old implementation used quaternion multiplication to transport
the argument to the Lie algebra, and use the standard formula there.
However, one may as well use the formula for the unit sphere in R^4,
which is much simpler, and also faster.  I fact, this patch gives
a few percent of speedup for the overall code.

[[Imported from SVN: r9562]]