Revert patch 9558
We normalize unit vectors again in the constructor and the assignment operator.
This makes sure we never drift away from the unit sphere, and it also allows
us to init unit spheres with any value in R^n and be sure we obtain a unit
vector.  This makes the test pass again.  Leaving the projection out didn't
really make a measurable difference anyway.

[[Imported from SVN: r9574]]