Implement vector-mode AD for the Hesse matrix
This almost halves the assembly time for the Hesse matrix.  Nice!
Nevertheless I still keep the scalar-mode code around, hidden
behind and #ifdef.  Higher-order vector-mode in ADOL-C appears
to be not quite as reliable as the rest, so I want to be able to
easily compare results with the scalar code.

Also, unlike the scalar code, the vector-mode one again computes
the full Hessian in the surrounding Euclidean space.  Only afterwards,
that full Hessian is multiplied by the tangent space basis vectors
to obtain a Hesse matrix on the tangent space.  To obtain better
efficiency one could have ADOL-C compute the Hessian in the tangent
directions directly.  However, ADOL-C sometimes segfaults... when
doing that.

[[Imported from SVN: r9815]]