Remove the code sharing local sizes from the constructor
That code was buggy, and we only didn't see that because the wrong sizes
were always overwritten by the correct ones before being used.
(That's so as long as you call a reducing method before 'scatter').

Removing the code leaves the situation as bad as before, but it
allows us to get rid of the nOwnedLocalEntity() method from the global
mapper.

[[Imported from SVN: r9926]]