Interpolation functions that works by retracting onto a tangent space
This is the interpolation method used by Ingo Münch and Wolfgang Müller for
Cosserat materials.  My implementation is only able to do all that I needed
for the illustrations in the GAMM Rundbrief article.  I'm sure there'll be
more fixes needed if you actually want to do finite elements with this.

[[Imported from SVN: r9940]]