[bugfix] The matrix communicator actually needs to different GridView objects
Normally, the two different level grid views.  Previously we used the same one
to compute the localToGlobal renumbering for boths rows and columns.  Of course
that screws up the numbering.

I am not quite sure when that got introduced.  I may have gotten in fairly recently,
after the upstreaming of the GlobalIndexSet classs.

[[Imported from SVN: r9962]]