Have an extra multigrid level when using a second- or third-order FE space
Previously, the number of multigrid transfer operators was always equal to
the number of grid levels minus one.  That worked out correctly when doing
first-order FE spaces.  However, for higher-order spaces, the up-most transfer
operator transfers from the higher-order space to a first-order space
_on the same grid level_.  But since the total number of transfer operators
was grid levels minus one, this implied that the coarsest grid leve 
was never used, and the coarse grid solver would operate on the second-coarsest
grid.

This patch adds the extra transfer operator in this case.  That means we now
call the coarse grid solver on smaller problems, which makes a noticeable
run-time difference.

[[Imported from SVN: r9964]]