Replace PkToP1MGTransfer with BasisInterpolationMatrixAssembler from dune-fufem
There is no need to have our own implementation for this, we had it purely for
historical reasons.

Also, this change fixes a subtle bug:  Our own implementation used AlienElementLocalBasis,
which returns zero if the LocalBasis is evaluated outside of the reference element.
However, this happened sometimes due to rounding error on grids with very small quadrilateral
elements.

[[Imported from SVN: r9991]]