Make the DR variable smaller
The DR variable (the derivative of the microrotation R wrt space) had dimensions 3x3x3.
This is two big in a shell context, because there are actually only two space dimensions.
Hence the correct size is 3x3xgridDim.

The results without this patch were not wrong, though, only a bit wasteful.