Use Higham's method instead of the SVD to compute the polar decomposition
The method using the SVD works in principle.  However, it does not play nicely
with automatic differentiation, because singular values are not always differentiable.

The method proposed by Higham to compute the polar factor, on the other hand,
is a Newton method at heart.  Therefore, it can be used together with AD without
problems.