Fix one issue that prevented the code to compile with dimworld!=3
I want this file with dimworld!=3 grids, even though it only makes sense
to use it if dimworld==3.

With this patch the file still does not compile.  However, the only remaining
problem is the hacky computation of the surface normals, which needs to be
rewritten in full generality soon anyway.