Store the local energy in a pointer to the base class, rather than by value
This allows to select the energy at run-time, but this is not what we do.
It also allows to give extra parameters to the nonplanar Cosserat energy,
and that is what we need: in a subsequent patch the main program will compute
a surface normal field and pass it to the Nonplanar Cosserat energy.