Make output logging work in the Python testsuite
By using the same mechanism as in the shell script:  piping everything through
the 'tee' program.  This may not be the most elegant way from a Python standpoint,
but it seems to work.

Since now my Python testsuite seems to do everything the shell version did,
this commit also removes the shell version.