Use a true second-order function as the base of the second-order vector field
It seemed strange to me to demonstrate a second-order tangent field on a
first-order base function.  People may think that this is a necessary restriction.

Also, this patch includes a bugfix:  The basis of the tangent space was hard-wired
to (0,1,0) and (0,0,1).  But this only works if we are at the tangent space at (1,0,0).