Update to pre-2.6 dune-solvers interface
A lot of raw pointers have been replaced by std::shared_ptr.

A lot of data members are inaccessible now, and getters have
to be used.

A small patch is still needed for dune-solvers: The iterationStep_
member is still needed to be public, because the new getter only
grants access to the object, not the pointer to it.