Radical modernization of cosseratenergytest.cc
In particular, this patch removes tests for the gradient
of the Cosserat energy and related things.  The
CosseratEnergyLocalStiffness class stopped computing
these things back in 2014 (in 63e42ece72da4950160f61cdcc83eb38bea1d3c2 )
when it was decided that algorithmic differentiation was
the way to go.

With this patch, the test builds again, but it doesn't run yet.
I am afraid that fixing that may require a larger investment
of time...