Allow to use surface grids
In this case, two different refinements of the grids are not
necessarily geometrically identical.  This means we cannot
use hierarchic search with the global position of a quadrature
point, and expect to find the point on the other grid as well.

Instead, we go done to the coarsest grid, and make the transfer
to the other grid there.