Merge mixed-cosserat-continuum.cc into cosserat-continuum.cc
These two files are very similar, and merging them will reduce
maintenance effort in the future.  As of this patch, there is
now quite a number of preprocessor directives in the new merged
cosserat-continuum.cc.  That is not pretty, but the new state
is still much easier to maintain than two completely separate
files.  What is more, I hope to be able to reduce some of that
preprocessor trickery by smarter use of the dune-functions
basis trees.