Make dune-parmg an optional dependency of dune-gfe
dune-gfe contains a bit of home-grown infrastructure for distributed
assembly of the tangent stiffness matrix.  This will now be removed
in favour of dune-parmg code, for better maintainability.  Also,
eventually we want the parallel solver of dune-parmg in addition to
the parallel assembly.

The new dependency is optional, to cater for people without interest
in distributed computing.