Reimplement EmbeddedGlobalGFEFunctions in dune-functions style
The old style has been deprecated for a long time, and is not
supported by dune-fufem anymore.

(In particular, the code discretizationerror.hh now requires
functions to satisfy the dune-functions function interface.)

In the wake of this, compute-disc-error.cc gets a number
of changes.  Let's hope I didn't screw up anything.