Do not interpolate vector-valued data using scalar bases
This used to work A Long Time Ago, but the feature was removed
eventually.  Unfortunately, the code still used to build,
but failed to produce the correct result.  More recently,
some other commit in dune-functions apparently now leads to
a build failure when trying to call 'interpolate' for a
vector-valued function and a scalar basis.  This is nice!
It allows to catch errors earlier.