Test ADOL-C scalar mode with first-order GFEs
Currently this test fails on the CI system with an obscure message
about not being able to write the tape to a file.  Maybe reducing
the complexity of the functional (by reducing the FE order) will
get us around that.  In any case I think testing with first-order FE
is just as thorough as with second-order ones.