Don't include cosseratenergystiffness.hh from nonplanarcosseratshellenergy.hh
It does not use the main class CosseratEnergyLocalStiffness,
but only the small helper class to access local finite elements
from different types of basis trees.  I am not very happy about
this helper class, but let's be pragmatic and simply copy it
to nonplanarcosseratshellenergy.hh (with a new, unique name).